Add StorageService implementation and corresponding tests
A new class, StorageServiceImpl, which implements the StorageService interface, has been added. Also, a series of tests for new methods in the StorageServiceImplTest class have been created. Method return types and parameter types have been updated in the StorageService interface from int to long for quantity representation. Closes #37

+++ eve-tracker-web/src/main/java/squareonex/evetracker/services/StorageService.java
@@ -8,15 +8,15 @@ public interface StorageService {
 
     // Adds a specific quantity of an item to the storage
     // Returns the new total quantity of this item in the storage
-    int add(Item item, int quantityToAdd);
+    long add(Item item, long quantityToAdd);
 
     // Removes a specific quantity of an item from the storage
     // Returns the remaining quantity of this item in the storage, or an empty Optional if the item does not exist in the storage
-    Optional<Integer> remove(Item item, long quantityToRemove);
+    Optional<Long> remove(Item item, long quantityToRemove);
 
     // Checks the quantity of an item in the storage
     // Returns the quantity of this item in the storage, or an empty Optional if the item does not exist in the storage
-    Optional<Integer> check(Item item);
+    Optional<Long> check(Item item);
     // Checks if the storage contains the specified quantity of an item
     // Returns true if the storage contains at least the specified quantity of the item, or false otherwise
     boolean isAvailable(Item item, long quantity);
